requirements.txt
tensorflow
opencv-python
gradio
numpy
Pillow
